Fixes #1426 "Extend" merge behavior for molecules
